+++ src/components/main/actionsPanel/navigatorPanel/helpers/getFileNameFromPath.ts
@@ -1,0 +1,5 @@
+import { TFile } from "@_types/main";
+
+export const getFileNameFromPath = (file: TFile) => {
+  return file.uid.split("/")[file.uid.split("/").length - 1];
+};
