+++ src/components/main/actionsPanel/navigatorPanel/helpers/index.ts
@@ -1,0 +1,5 @@
+export { isSelected } from "./isSelected";
+export { isHomeIcon } from "./isHomeIcon";
+export { getFileNameFromPath } from "./getFileNameFromPath";
+export { setWorkspaceFavicon } from "./setWorkspaceFavicon";
+export { selectFirstNode } from "./selectFirstNode";
