+++ src/components/main/actionsPanel/navigatorPanel/helpers/isSelected.ts
@@ -1,0 +1,9 @@
+import { TProject } from "@_types/main";
+
+export const isSelected = (_project: TProject, project: TProject) => {
+  return _project.context === project.context &&
+    _project.name === project.name &&
+    _project.handler === project.handler
+    ? "selected"
+    : "";
+};
